requirements.txt
absl-py==1.2.0
asgiref==3.4.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
deepface==0.0.75
Django==4.0.1
django-appconf==1.0.5
django-cors-headers==3.10.1
django-environ==0.10.0
django-imagekit==4.1.0
django-quill-editor==0.1.36
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
flatbuffers==2.0.7
gast==0.4.0
gdown==4.5.1
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mtcnn==0.1.1
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
pilkit==2.0
Pillow==9.0.0
protobuf==3.19.5
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
textdistance==4.3.0
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2021.5
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.15.0
